PodsList: default to expand all
Phase is confusing and easily confused with table printer concluded
status for beginners, but i still want something easy to implement that
can represent the state of the whole pod here. It is at least useful to
distiguish if it is a finished one-time pod like from job.

also there are typically not much pods in a namespace.

this make the default behavior perhaps friendlier.